added route redirects, started fixing home page styles & building layout for hero component
