+++ package.json
@@ -22,7 +22,7 @@
   "dependencies": {
     "@auth/prisma-adapter": "^1.0.0",
     "@mui/material": "^5.13.5",
-    "@prisma/client": "^4.16.1",
+    "@prisma/client": "^4.16.2",
     "@types/bcrypt": "^5.0.0",
     "@types/react": "18.2.12",
     "@types/react-dom": "18.2.5",
