+++ src/app/dashboard/layout.tsx
@@ -1,0 +1,7 @@
+export default function DashboardWrapperLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <div>{children}</div>;
+}
