+++ src/components/landing/hero.module.scss
@@ -1,0 +1,5 @@
+.container {
+  border: 2px solid red;
+  display: flex;
+  align-items: center;
+}
