+++ src/components/landing/hero.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import styles from "./hero.module.scss";
+
+const hero = () => {
+  return (
+    <div className={styles.container}>
+      <h1>Hero</h1>
+    </div>
+  );
+};
+
+export default hero;
